Revised caching functionality
Caching now monitors changes in file content
of the original files and parameter changes
in order to decide whether to reload a cache file.
Consequently, the attributes datatags and overwrite
become obsolete.
A deprecation warning is issued if datatags or overwrite
are set. They will be removed in the future.