condition, resolution and order reuse changed
Before, these parameters where stored and reused from the cachefile
directly.
However, it might lead to misleading effects if the cached parameter
does not agree with the supplied parameter in the constructor. Then the
parameter will be overwritten by the cached version.
To fix this, the cached parameter will be ignored. It will however still
be written to the cache file in  order to have the cache files backwards
compatible with the current version.