fixed bioseq padding store_whole_genome=True
By default the genomic array was zero-padded when intervals where
reaching over the chromosome starts or ends.
However, since the values in the genomic array are interpreted directly
as indices for the one hot encoding, the padded values are interpreted
as 'A's.
In this fix, a padding value can be passed to the constructors of
genomic arrays allowing to set padding values that are interpreted as
'N's in the DNA sequence rather than 'A's.