Fixed #8
The solution involved more carefully parsing the genomic interval from a
string like 'chr:start-end'. The string splitting approach did not work
if another '-' character was contained in the string, which came from
the minus sign. This was fixed.
In addition, to make zero-padding work, it was necessary to add 'N'
before or after the original sequence during loading to fill up the
remaining possitions.