Added lazy loading mechanism.
the genome size from the fasta reference genome is derived by loading
and determining the sequence lengths. however, the gsize is only needed
for specifying the dataset size upon construction. when fetchin cached
data, the gsize is not needed. lazy loading takes that into acount which
saves time to reload the Bioseq dataset from reference.